restructering repo + block structure introduction
