+++ src/f3dasm/__init__.py
@@ -15,8 +15,8 @@ optimization, sampling, and simulation.
 
 from .__version__ import __version__
 from ._src._argparser import HPC_JOBID
-from ._src.experimentdata.experimentdata import ExperimentData
-from ._src.experimentdata.experimentsample import ExperimentSample
+from ._src.experimentdata import ExperimentData
+from ._src.experimentsample import ExperimentSample
 from ._src.logger import DistributedFileHandler, logger
 
 #                                                        Authorship and Credits
