+++ src/f3dasm/_src/design/__init__.py
@@ -7,6 +7,7 @@ Design-of-experiments (DOE) module for the f3dasm package.
 
 # Local
 from .domain import Domain, _domain_factory
+from .samplers import _sampler_factory
 
 #                                                          Authorship & Credits
 # =============================================================================
@@ -19,4 +20,5 @@ __status__ = 'Stable'
 
 __all__ = ['Domain',
            '_domain_factory',
+           '_sampler_factory'
            ]
