+++ src/f3dasm/_src/utils.py
@@ -1,0 +1,37 @@
+def number_of_updates(iterations: int, population: int):
+    """Calculate number of update steps to acquire the
+     correct number of iterations
+
+    Parameters
+    ----------
+    iterations
+        number of desired iteration steps
+    population
+        the population size of the optimizer
+
+    Returns
+    -------
+        number of consecutive update steps
+    """
+    return iterations // population + (iterations % population > 0)
+
+
+def number_of_overiterations(iterations: int, population: int) -> int:
+    """Calculate the number of iterations that are over the iteration limit
+
+    Parameters
+    ----------
+    iterations
+        number of desired iteration steos
+    population
+        the population size of the optimizer
+
+    Returns
+    -------
+        number of iterations that are over the limit
+    """
+    overiterations: int = iterations % population
+    if overiterations == 0:
+        return overiterations
+    else:
+        return population - overiterations
