+++ src/f3dasm/optimization.py
@@ -5,8 +5,8 @@ Module for optimization
 # =============================================================================
 
 # Local
+from ._src.core import Optimizer
 from ._src.optimization import cg, lbfgsb, nelder_mead, random_search
-from ._src.optimization.optimizer import Optimizer
 from ._src.optimization.optimizer_factory import available_optimizers
 
 #                                                          Authorship & Credits
