+++ tests/design/test_parameters.py
@@ -2,9 +2,10 @@ import pickle
 
 import pytest
 
-from f3dasm._src.design.parameter import (  # Replace with actual module name
-    CategoricalParameter, ConstantParameter, ContinuousParameter,
-    DiscreteParameter, Parameter)
+from f3dasm._src.design.parameter import (CategoricalParameter,
+                                          ConstantParameter,
+                                          ContinuousParameter,
+                                          DiscreteParameter, Parameter)
 
 pytestmark = pytest.mark.smoke
