+++ tests/functions/test_scaling_and_offset.py
@@ -24,7 +24,7 @@ def test_offset(function: Function, seed: int):
         scale_bounds=domain.get_bounds(),
     )
 
-    func.init(data=ExperimentData(domain=domain))
+    func.arm(data=ExperimentData(domain=domain))
 
     xmin = func._get_global_minimum_for_offset_calculation()
 
@@ -56,7 +56,7 @@ def test_check_global_minimum(function: str):
     seed = 42
     func = _datagenerator_factory(
         data_generator=function, seed=seed)
-    func.init(data=ExperimentData(domain=domain))
+    func.arm(data=ExperimentData(domain=domain))
     _ = func.get_global_minimum(dim)
 
 
@@ -90,7 +90,7 @@ def test_scaling_1(function: str, scale_bounds_list: List[float]):
         [-1.0, 1.0], (dim, 1)), dimensionality=dim)
     func = _datagenerator_factory(
         data_generator=function, seed=seed, scale_bounds=scale_bounds)
-    func.init(data=ExperimentData(domain=domain))
+    func.arm(data=ExperimentData(domain=domain))
     x = np.random.uniform(
         low=scale_bounds[0, 0], high=scale_bounds[0, 1],
         size=(1, func.dimensionality))
